Secure Reversible Data Hiding in Encrypted Images based on Classification Encryption Difference
This paper introduces an algorithm to improve the
security, efficiency, and embedding capacity of reversible data
hiding in encrypted images (RDH-EI). It is based on classification
encryption difference and adaptive fixed-length coding. Firstly,
the prediction error image is obtained, the difference with a bin
value greater than the encryption threshold in the difference
histogram is found, and it is further modified to obtain the embedding threshold range. Then, under the condition of ensuring
that the difference inside and outside the embedding threshold
range is not confused, the difference within the threshold is only
scrambled, and the difference outside the threshold is scrambled
and mod encrypted. After obtaining the encrypted image, an
adaptive difference fixed-length coding method is proposed to
encode and compress the differences within the threshold. The
secret data is embedded in the multiple most significant bits
of the encoded difference. Experimental results show that the
embedding capacity of the proposed algorithm is improved
compared with the state-of-the-art algorithm.